line/styled: Proposal to improve dot / line alignment further
The upstreamed dotted line offset is correct, but it doesn't always
change when the line thickness changes, which is ugly.

This commit modifies the offset to change when the line thickness
changes, but this introduces the weird issue that the last dot moves...
So this commit also changes the extend_line function.

This is not ideal for width = 2, but the upstreamed version isn't
either.

NOTE (1 year later) : this is a very small change. Yes, it seems to
align width changes and end dot position changes. The upstreamed version
and this version both are bad for the 2px line.